+++ changelog.md
@@ -23,3 +23,4 @@
 * Use Django REST framework status code constants instead of integers.
 * Added long text field.
 * Fixed not refreshing token bug and improved authentication a little bit.
+* Introduced copy, paste and delete functionality of selected fields.